krux-stdlib-monitoring -> krux-stdlib-main
The old name wasn't accurate since it was really "everything except
logger config", and we no longer need krux-stdlib-logging.
