+++ krux-stdlib-main/resources/log4j2.properties
