+++ krux-stdlib-main/src/main/java/com/krux/stdlib/KruxHelpFormatter.java
