+++ krux-stdlib-main/src/main/java/com/krux/stdlib/shutdown/Shutdownable.java
