+++ krux-stdlib-main/src/main/java/com/krux/stdlib/statsd/StatsdStatType.java
