+++ krux-stdlib-main/src/main/java/com/krux/stdlib/status/AppState.java
