+++ krux-stdlib-main/src/main/java/com/krux/stdlib/status/StatusBean.java
