+++ krux-stdlib-main/src/main/java/com/krux/stdlib/utils/ExternalProperties.java
