+++ krux-stdlib-main/src/test/java/com/krux/stdlib/utils/ExternalPropertiesTest.java
