+++ krux-stdlib-main/src/test/resources/test_properties.properties
