+++ pom.xml
@@ -10,11 +10,12 @@
   <url>https://github.com/krux/java-krux-stdlib</url>
 
   <modules>
-    <module>krux-stdlib-monitoring</module>
+    <module>krux-stdlib-main</module>
     <module>krux-stdlib-examples</module>
   </modules>
 
   <properties>
+    <assertj.version>3.13.2</assertj.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <commons-lang.version>3.3.2</commons-lang.version>
     <config.version>4.8</config.version>
@@ -42,9 +43,9 @@
     <dependencies>
       <!-- testing -->
       <dependency>
-	<groupId>org.junit.jupiter</groupId>
+	      <groupId>org.junit.jupiter</groupId>
         <artifactId>junit-jupiter-api</artifactId>
-	<version>${junit.jupiter.version}</version>
+	      <version>${junit.jupiter.version}</version>
         <scope>compile</scope>
       </dependency>
       <dependency>
@@ -63,6 +64,12 @@
         <groupId>junit</groupId>
         <artifactId>junit</artifactId>
         <version>${junit.version}</version>
+        <scope>test</scope>
+      </dependency>
+      <dependency>
+        <groupId>org.assertj</groupId>
+        <artifactId>assertj-core</artifactId>
+        <version>${assertj.version}</version>
         <scope>test</scope>
       </dependency>
 
@@ -113,7 +120,7 @@
 
       <dependency>
         <groupId>com.krux</groupId>
-        <artifactId>krux-stdlib-monitoring</artifactId>
+        <artifactId>krux-stdlib-main</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
@@ -134,7 +141,6 @@
           <source>${maven.compiler.source}</source>
           <target>${maven.compiler.target}</target>
           <compilerArgs>
-            <arg>-verbose</arg>
             <arg>-Xlint:all</arg>
           </compilerArgs>
         </configuration>
